feat: ManagerCustomerCommandServiceTest 모킹테스트로 변경
